+++ hutool-crypto/src/main/java/cn/hutool/crypto/digest/mac/BCHMacEngine.java
@@ -14,9 +14,7 @@ import org.bouncycastle.crypto.params.ParametersWithIV;
  * @author Looly
  * @since 4.5.13
  */
-public class BCHMacEngine implements MacEngine {
-
-	private Mac mac;
+public class BCHMacEngine extends BCMacEngine {
 
 	// ------------------------------------------------------------------------------------------- Constructor start
 
@@ -51,7 +49,18 @@ public class BCHMacEngine implements MacEngine {
 	 * @since 4.5.13
 	 */
 	public BCHMacEngine(Digest digest, CipherParameters params) {
-		init(digest, params);
+		this(new HMac(digest), params);
+	}
+
+	/**
+	 * 构造
+	 *
+	 * @param mac {@link HMac}
+	 * @param params 参数，例如密钥可以用{@link KeyParameter}
+	 * @since 5.8.0
+	 */
+	public BCHMacEngine(HMac mac, CipherParameters params) {
+		super(mac, params);
 	}
 	// ------------------------------------------------------------------------------------------- Constructor end
 
@@ -61,46 +70,9 @@ public class BCHMacEngine implements MacEngine {
 	 * @param digest 摘要算法
 	 * @param params 参数，例如密钥可以用{@link KeyParameter}
 	 * @return this
+	 * @see #init(Mac, CipherParameters)
 	 */
 	public BCHMacEngine init(Digest digest, CipherParameters params) {
-		mac = new HMac(digest);
-		mac.init(params);
-		return this;
-	}
-
-	/**
-	 * 获得 {@link Mac}
-	 *
-	 * @return {@link Mac}
-	 */
-	public Mac getMac() {
-		return mac;
-	}
-
-	@Override
-	public void update(byte[] in, int inOff, int len) {
-		this.mac.update(in, inOff, len);
-	}
-
-	@Override
-	public byte[] doFinal() {
-		final byte[] result = new byte[getMacLength()];
-		this.mac.doFinal(result, 0);
-		return result;
-	}
-
-	@Override
-	public void reset() {
-		this.mac.reset();
-	}
-
-	@Override
-	public int getMacLength() {
-		return mac.getMacSize();
-	}
-
-	@Override
-	public String getAlgorithm() {
-		return this.mac.getAlgorithmName();
+		return (BCHMacEngine) init(new HMac(digest), params);
 	}
 }
